requirements.txt
ultralytics>=8.4,<9
pillow>=10
